Improve the code
