+++ keyboards/miiiw/blackio83/halconf.h
@@ -30,4 +30,3 @@
 #define USB_USE_WAIT TRUE
 
 #include_next <halconf.h>
-
